Restructure ParsedHedSubstring family and fix definition issues
