+++ tests/event.spec.js
@@ -2,7 +2,7 @@ const assert = require('chai').assert
 const hed = require('../validator/event')
 const { buildSchemas } = require('../validator/schema/init')
 const { parseHedString } = require('../validator/parser/main')
-const { ParsedHedTag } = require('../validator/parser/types')
+const { ParsedHedTag } = require('../validator/parser/parsedHedTag')
 const { HedValidator, Hed2Validator, Hed3Validator } = require('../validator/event')
 const { generateIssue } = require('../common/issues/issues')
 const converterGenerateIssue = require('../converter/issues')
