+++ tests/stringParser.spec.js
@@ -2,7 +2,8 @@ const assert = require('chai').assert
 const { Schemas } = require('../common/schema')
 const { parseHedString } = require('../validator/parser/main')
 const splitHedString = require('../validator/parser/splitHedString')
-const { ParsedHedTag, ParsedHedSubstring } = require('../validator/parser/types')
+const ParsedHedSubstring = require('../validator/parser/parsedHedSubstring')
+const { ParsedHedTag } = require('../validator/parser/parsedHedTag')
 const { generateIssue } = require('../common/issues/issues')
 const { SchemaSpec, SchemasSpec } = require('../common/schema/types')
 const converterGenerateIssue = require('../converter/issues')
